# backend/services/api-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Music Room Gateway API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development gateway
paths:
...
  /users/me:
    get:
      summary: Get current user's profile
      tags: [users]
      security:
        - bearerAuth: []   # Фронт/мобилка должны передавать токен(обязателен JWT-эндпоинт), иначе API Gateway ответит 401
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update current user's profile
      tags: [users]
      security:
        - bearerAuth: []   # обязательный JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      summary: Get public view of a user profile by user ID
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID (same as auth userId)
      responses:
        '200':
          description: Public view of user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUserProfile'
        '404':
          description: User or profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # --------------------
  # PLAYLISTS
  # --------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UserVisibility:
      type: string
      enum: [public, friends, private]
      default: public
    MusicPreferences:
      type: object
      properties:
        genres:
          type: array
          items:
            type: string
        artists:
          type: array
          items:
            type: string
        moods:
          type: array
          items:
            type: string
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: Profile ID
        userId:
          type: string
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        visibility:
          $ref: '#/components/schemas/UserVisibility'
        preferences:
          $ref: '#/components/schemas/MusicPreferences'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, displayName, visibility, preferences, createdAt, updatedAt]
    PublicUserProfile:
      type: object
      properties:
        id:
          type: string
          description: User ID (owner of the profile)
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        visibility:
          $ref: '#/components/schemas/UserVisibility'
        preferences:
          $ref: '#/components/schemas/MusicPreferences'
      required: [id, displayName, visibility]
    UpdateUserProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
          format: uri
        visibility:
          $ref: '#/components/schemas/UserVisibility'
        preferences:
          $ref: '#/components/schemas/MusicPreferences'

    # ---------- PLAYLISTS ----------
...